Fixed critical bug in IOSet and added Bools
The bug in IOSet created EqRepr that pointed on classes that should not
exists anymore (becuase of union). Sometimes we wrote data to the nodes
(actually everywhere) insteed of making sure that we wrote to the root.
That created duplicates of some classes.

I also made our patterns a bit more dynamical like:
    identity op v = forall1 $ \x -> (x `op` v) ~> x
    commute op    = forall2 $ \x y -> (x `op` y) ~> (y `op` x)

we also have equality
    forall1 $ \x -> x `peq` x ~> pbool True

which solves
    from:
        ((LBool True == LBool False) `or` (LInteger 2 == (LInteger 2 * LInteger
1))) @ #0
    to:
        (1,LBool True)