Added a parser and a translation
We translate Expr to Opt EqRepr, this is similar to what addExpr
did in Expr.hs But now we don't add variables that are bound. Instead
it point to the class that the expression of the var is in.

A new function exists now in TestExpr to test it, testFileExpr which
takes a filePath and parses it and then run the optimisation on it :)