addTerm now checks against all other terms
This takes more time, but I have implemented a smarter dirty-check
to compensate for that.

Now buildPattern in Rules.hs also returns a bool if the current cls
has changed. Previously we checked if we didn't found patterns, but
now we check if the pattern we found made any change to the class.

I thought this would allow us to set the iteration to crazy high
since nothing will change after it is saturated but there seems to
be something that makes it do work (I changed the iteration count
from 12 to 12000 and it took about half a minute (haven't timed so
this is based on feeling).