Started to implement a better pattern matcher.
It still doesn't use all shortcuts it should be able to, but I think
it exposes more oppurtunity for optimizing it. It should for example
be simpler to say that these rules have already been used also smarter
data structure can save work (I have tried to make it easy to change).

So the general idea is that we try to match all the rules at the same
time. So we now have a type ToCheck which is sort of a continuation
of all the patterns and the meta data for matching and continue for
all the rules.

So Done is all the matches that has succeded (this could be filtered
by all we have found previous). Also the way we insert the ToCheck
could be smarter and merge the Checks if the classes they apply to
are similar, that would save us time when going through that class.