Added graphviz support
With the new -d flag you can specify a filepath to save a dot file in
for example:

$ PureEqSat -f examples/ex5.pes -d ex5.dot
$ dot -Tpng -O ex5.dot

Will create a ex5.png that is a representation of the equality graph.
The graph itself is of the classes and which classes that they depend
on.

With this I found out that ex6.pes has a cycle which could explain why
it creates alot of classes.