It is faster, and more optimizations is available.
Two things have changed:

1) if we want to know which class a + b is in we don't wanna look
in the classes that depend on either a or b but thoose that depend
on both. This speedsup the process of actually carrying out the
applications of the rules.

2) a smarter way of looking for patternmatching, my new approach of
trying to match against as many rules at the same time had a small
simplification. I had at the designstage thought about this but later
forgot: when inserting new things ToCheck I just added them to the list
but if we join all the Checks that correspond to that class we will save
a lot of work. This made things quite faster (even though I do this
rather stupid at the moment). A better datatype might yield even better
performance.

Still we have a problem that we create quite a lot of classes and
we need to look into why.